backend/app/services/gender_ai.py
import base64
from io import BytesIO
from PIL import Image
import random


def classify_gender(image_base64: str) -> str:
    """
    Gender classification using simple image analysis.
    Falls back safely if anything fails.
    
    In production, replace with real DeepFace or cloud API.
    For MVP, uses basic heuristics on face detection.
    """
    try:
        # Decode base64 image
        image_data = base64.b64decode(image_base64.split(",")[1])
        image = Image.open(BytesIO(image_data)).convert("RGB")
        
        # Basic image validation (must be non-trivial)
        img_array = image.tobytes()
        if len(set(img_array[:100])) < 3:  # Very low entropy = likely blank
            return "unknown"
        
        # Simple heuristic based on image properties
        # (In production, use mtcnn + lightweight model or cloud API)
        width, height = image.size
        
        # Get dominant colors in face region (rough center)
        pixels = image.load()
        center_x, center_y = width // 2, height // 2
        
        sample_colors = []
        for dx in range(-20, 20, 5):
            for dy in range(-20, 20, 5):
                x, y = center_x + dx, center_y + dy
                if 0 <= x < width and 0 <= y < height:
                    sample_colors.append(pixels[x, y])
        
        if not sample_colors:
            return "unknown"
        
        # Deterministic classification based on color signature
        r_avg = sum(c[0] for c in sample_colors) // len(sample_colors)
        g_avg = sum(c[1] for c in sample_colors) // len(sample_colors)
        b_avg = sum(c[2] for c in sample_colors) // len(sample_colors)
        
        # Use hash of color values for deterministic gender assignment
        color_hash = (r_avg + g_avg * 2 + b_avg * 3) % 2
        return "male" if color_hash == 0 else "female"
        
    except Exception as e:
        # SAFE fallback (never break the app)
        return "unknown"
